requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Redis & Caching
redis==5.2.0
hiredis==3.0.0

# Task Queue
celery==5.4.0
flower==2.0.1

# Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.27.1
playwright==1.49.0
fake-useragent==1.5.1

# Async Support
httpx==0.28.0
aiofiles==24.1.0
aiohttp==3.11.10

# Data Processing
pandas==2.2.3
numpy==2.1.3
python-dateutil==2.9.0

# LLM Integration
google-generativeai==0.8.3
langchain==0.3.9
langchain-google-genai==2.0.5

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
email-validator==2.2.0

# Monitoring & Logging
loguru==0.7.3
prometheus-client==0.21.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.0

# Type Checking
mypy==1.13.0
types-requests==2.32.0.20241016